refactor to single file component
