+++ .gitignore
@@ -2,3 +2,5 @@ package-lock.json
 gcloudstoragekey.json
 .env
 node_modules
+dist
+.cache
