--- client/js/components/myarticles.js
@@ -1,17 +1,0 @@
-Vue.component("myarticles-page", {
-  data() {
-    return {};
-  },
-  created() {},
-  mounted() {},
-  methods: {
-    _getUserArticles() {}
-  },
-  template: `
-  <div class="p-5">
-    <div class="border">
-      MY ARTICLES PAGE COMPONENT
-    </div>
-  </div>
-  `
-});
