--- client/js/main.js
@@ -1,13 +1,0 @@
-$(document).ready(function() {
-  $("a").click(function(event) {
-    event.preventDefault();
-  });
-
-  $("button").click(function(event) {
-    event.preventDefault();
-  });
-  
-  $("form").on("submit", function(event) {
-    event.preventDefault();
-  });
-});
