--- client/js/vue.js
@@ -1,384 +1,0 @@
-const serverURL = "http://localhost:3000";
-const ax = axios.create({
-  baseURL: serverURL,
-  headers: {
-    token: localStorage.getItem("token")
-  }
-});
-
-const vue = new Vue({
-  el: "#app",
-  data: {
-    pages: [
-      "articles",
-      "detailarticle",
-      "myarticles",
-      "newarticle",
-      "editarticle",
-      "register",
-      "login",
-      "setting"
-    ],
-    isOnPage: undefined,
-    isLogin: undefined,
-    articles: [],
-    tags: [],
-    inputLoginRegister: {
-      name: "",
-      email: "",
-      password: ""
-    },
-    inputArticle: {
-      title: "",
-      content: "",
-      rawHTML: "",
-      picture: "",
-      tags: []
-    },
-    inputFilter: "",
-    inputTag: "",
-    editor: ""
-  },
-
-  created() {},
-
-  mounted() {
-    this.checkLogin();
-    if (localStorage.getItem("isOnPage")) {
-      this.isOnPage = localStorage.getItem("isOnPage");
-    }
-  },
-
-  methods: {
-    // RESET THE INPUT FUNCTION
-    r_inputLoginRegister() {
-      this.inputLoginRegister = {
-        name: "",
-        email: "",
-        password: ""
-      };
-    },
-    r_inputLoginRegister_afterregister() {
-      this.inputLoginRegister.name = "";
-      this.inputLoginRegister.password = "";
-      swal("Account Created", "Successfully created an account", "success");
-    },
-    r_inputArticle() {
-      this.inputArticle = {
-        title: "",
-        content: "",
-        snippet: "",
-        rawHTML: "",
-        img: "",
-        tags: []
-      };
-    },
-
-    // INITIAL CHECK
-    checkLogin() {
-      console.log("check login");
-      if (localStorage.getItem("token")) {
-        this.isLogin = true;
-        if (localStorage.getItem("isOnPage")) {
-          this.isOnPage = localStorage.getItem("isOnPage");
-          this.loadGAPI();
-        } else {
-          console.log(localStorage.getItem("isOnPage"));
-
-          this.page_articles();
-        }
-      } else {
-        this.isLogin = false;
-        this.page_login();
-      }
-    },
-
-    selectArticlePic(event) {
-      this.inputArticle.picture = event.target.files[0];
-    },
-
-    // GOOGLE AUTH FUNCTION
-    loadGAPI() {
-      gapi.load("auth2", () => {
-        auth2 = gapi.auth2.init({
-          client_id:
-            "647848106811-cm0ck452dmhg5b8g38n0dh2bu8ovv7hh.apps.googleusercontent.com",
-          cookiepolicy: "single_host_origin"
-        });
-
-        if (this.isOnPage === "login") {
-          auth2.attachClickHandler(
-            "g-signin2",
-            {},
-            googleUser => {
-              document.getElementById("g-signin2").innerText =
-                "Signed in: " + googleUser.getBasicProfile().getName();
-            },
-            error => {
-              console.log(JSON.stringify(error, null, 2));
-            }
-          );
-        }
-        auth2.isSignedIn.listen(this.renderGoogleButton);
-      });
-
-      if (this.isOnPage === "login") {
-        this.renderGoogleButton();
-      }
-    },
-    renderGoogleButton() {
-      gapi.signin2.render("g-signin2", {
-        scope: "profile email",
-        longtitle: true,
-        theme: "light",
-        onsuccess: this.googlelogin
-      });
-    },
-    googlelogin(googleUser) {
-      console.log(googleUser);
-      const idToken = googleUser.getAuthResponse().id_token;
-      console.log(googleUser);
-      ax({
-        method: "POST",
-        url: "/users/logingoogle",
-        data: {
-          idToken
-        }
-      })
-        .then(({ data }) => {
-          localStorage.setItem("token", data.token);
-          localStorage.setItem("user", JSON.stringify(data.user));
-          swal(
-            "Logged In",
-            "If this is your first time log in, change your password by accessing Account tab above!",
-            "info"
-          );
-          this.checkLogin();
-        })
-        .catch(err => {
-          console.log(err, "error login google");
-          console.log(JSON.stringify(err, null, 2));
-        });
-    },
-
-    // CK EDITOR RELATED
-    load_editor() {
-      var toolbarOptions = [
-        ["bold", "italic", "underline", "strike"], // toggled buttons
-        ["blockquote", "code-block"],
-        [{ header: 1 }, { header: 2 }], // custom button values
-        [{ list: "ordered" }, { list: "bullet" }],
-        [{ script: "sub" }, { script: "super" }], // superscript/subscript
-        [{ indent: "-1" }, { indent: "+1" }], // outdent/indent
-        [{ direction: "rtl" }], // text direction
-        [{ size: ["small", false, "large", "huge"] }], // custom dropdown
-        [{ header: [1, 2, 3, 4, 5, 6, false] }],
-        [{ color: [] }, { background: [] }], // dropdown with defaults from theme
-        [{ font: [] }],
-        [{ align: [] }],
-        ["clean"] // remove formatting button
-      ];
-      if (document.querySelector("#editor")) {
-        var quill = new Quill("#editor", {
-          theme: "snow",
-          toolbar: toolbarOptions
-        });
-        this.editor = quill;
-        quill.on("text-change", function(delta, oldDelta, source) {
-          vue.inputArticle.content = quill.getText();
-          vue.inputArticle.rawHTML = quill.root.innerHTML;
-        });
-
-        if (this.inputArticle.rawHTML) {
-          quill.root.innerHTML = this.inputArticle.rawHTML;
-        }
-      }
-    },
-
-    // HELPERS FUNCTION
-    toTitleCase(str) {
-      let temp = str.split("");
-      temp[0] = temp[0].toUpperCase();
-      for (let i = 1; i < temp.length; i++) {
-        temp[i] = temp[i].toLowerCase();
-      }
-      str = temp.join("");
-      return str;
-    },
-
-    // ARTICLE RELATED FUNCTION
-    newArticle(status) {
-      let quill = this.editor;
-      let content = quill.getText();
-      let rawHTML = quill.root.innerHTML;
-      let { picture, titile, tags } = this.inputArticle;
-      let inputVal = {
-        title: this.inputArticle.title,
-        content,
-        rawHTML,
-        status,
-        tags,
-        picture
-      };
-      if (status.toLowerCase() === "post") {
-        inputVal.postedAt = new Date();
-      }
-
-      let errmsg = "";
-      if (inputVal.rawHTML == "<p><br></p>" || inputVal.content == "\n") {
-        errmsg += "Content of an article should not be empty\n";
-      }
-      if (inputVal.title == "") {
-        errmsg += "Title of an article should not be empty\n";
-      }
-      if (inputVal.tags == "") {
-        errmsg += "Article should have at least one tag\n";
-      }
-      if (inputVal.picture == "") {
-        errmsg += "Article should have a featured picture\n";
-      }
-
-      if (errmsg !== "") {
-        swal("Attention", errmsg, "info");
-      } else {
-        swal({
-          title: "Confirmation",
-          text: `${status} this article?`,
-          icon: "info",
-          buttons: true,
-          dangerMode: true
-        }).then(confirm => {
-          if (confirm) {
-            const blob = new Blob([inputVal.picture], {
-              type: inputVal.picture.type
-            });
-
-            const formdata = new FormData();
-            formdata.append("image", blob);
-
-            ax({
-              method: "POST",
-              url: "/uploadimg",
-              data: formdata,
-              headers: {
-                "Content-Type": "multipart/form-data",
-                token: localStorage.getItem("token")
-              }
-            })
-              .then(({ data }) => {
-                this.inputArticle.picture = data;
-                inputVal.picture = data;
-                return ax({
-                  method: "POST",
-                  url: "/articles",
-                  data: inputVal
-                });
-              })
-              .then(({ data }) => {
-                this.r_inputArticle();
-                quill.setText("");
-                $("#selectedPicture").val("");
-                swal("Yay", `Article ${this.toTitleCase(status)}ed`, "success");
-              })
-              .catch(err => {
-                swal("Sorry", "Problem occured, try again later", "error");
-                console.log(
-                  "error new article dengan upload gambar",
-                  JSON.stringify(err, null, 2)
-                );
-                console.log(err);
-              });
-          }
-        });
-      }
-    },
-    addTag() {
-      if (this.inputTag !== "") {
-        if (!this.inputArticle.tags) {
-          this.inputArticle.tags = [];
-        }
-        if (this.inputArticle.tags.length >= 5) {
-          swal("you can only add 5 tags per article");
-        } else {
-          let tags = this.inputArticle.tags;
-          let input = this.inputTag;
-          if (input.match(/^[A-Za-z]+$/)) {
-            if (tags.indexOf(input) === -1) {
-              if (input.length > 12 || input.length < 3) {
-                swal("Tags should consists of 3 - 12 characters");
-              } else {
-                this.inputArticle.tags.push(input.toLowerCase());
-              }
-            }
-          } else {
-            swal(
-              "Tags shouldnt contain any number, special character, and spaces!"
-            );
-          }
-        }
-        this.inputTag = "";
-      }
-    },
-    removeTag(tag) {
-      let tags = this.inputArticle.tags;
-      let index = tags.indexOf(tag);
-      if (index >= 0) {
-        tags.splice(index, 1);
-      }
-    },
-
-    // MOVING BETWEEN PAGE FUNCTION
-    page_articles() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[0];
-      localStorage.setItem("isOnPage", this.pages[0]);
-    },
-    page_detailarticles() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[1];
-      localStorage.setItem("isOnPage", this.pages[1]);
-    },
-    page_myarticles() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[2];
-      localStorage.setItem("isOnPage", this.pages[2]);
-    },
-    page_newarticle() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[3];
-      setTimeout(() => {
-        this.load_editor();
-      }, 100);
-    },
-    page_editarticle() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[4];
-      setTimeout(() => {
-        this.load_editor();
-      }, 100);
-    },
-    page_register() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[5];
-    },
-    page_login() {
-      this.r_inputLoginRegister();
-      this.isOnPage = this.pages[6];
-      setTimeout(() => {
-        this.loadGAPI();
-      }, 200);
-    },
-    page_login_afterregister(emit) {
-      this.r_inputLoginRegister_afterregister();
-      this.isOnPage = this.pages[6];
-      setTimeout(() => {
-        this.loadGAPI();
-      }, 200);
-    },
-    page_setting() {
-      this.r_inputLoginRegister();
-      localStorage.setItem("isOnPage", this.pages[7]);
-      this.isOnPage = this.pages[7];
-    }
-  }
-});
