+++ client/src/components/viewarticle.vue
@@ -1,0 +1,25 @@
+<template>
+  <div class="p-5">
+      VIEW ARTICLE
+  </div>
+</template>
+
+<script>
+export default {
+  props : ["selectedArticle"],
+  data() {
+    return {};
+  },
+  created () {},
+  mounted () {},
+  methods : {
+    like_unlike(){
+      
+    }
+  }
+};
+</script>
+
+<style scoped>
+</style>
+
