+++ client/src/main.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import App from "./app.vue";
+
+new Vue(App).$mount("#app");
